Simplify separated integration tests into standard unit tests
