+++ .gitattributes
@@ -1,6 +1,6 @@
 * text=auto eol=lf
 
 # Windows specific test files where we need CRLF endings
-crates/air/tests/fixtures/crlf/*.R text eol=crlf
+crates/air/fixtures/crlf/*.R text eol=crlf
 crates/air_r_formatter/tests/specs/r/crlf/*.R text eol=crlf
 crates/air_r_parser/tests/snapshots/ok/crlf/*.R text eol=crlf
