+++ crates/air/src/commands/format.rs
@@ -255,3 +255,50 @@ impl Display for FormatCommandError {
         }
     }
 }
+
+#[cfg(test)]
+mod test {
+    use crate::args::Args;
+    use crate::run;
+    use crate::status::ExitStatus;
+    use clap::Parser;
+    use std::path::Path;
+    use std::path::PathBuf;
+    use tempfile::TempDir;
+
+    fn path_fixtures() -> PathBuf {
+        Path::new(env!("CARGO_MANIFEST_DIR")).join("fixtures")
+    }
+
+    #[test]
+    fn default_options() -> anyhow::Result<()> {
+        let tempdir = TempDir::new()?;
+        let temppath = tempdir.path().join("test.R");
+        std::fs::write(
+            &temppath,
+            r#"
+1 + 1
+"#,
+        )?;
+
+        let args = Args::parse_from(["", "format", temppath.to_str().unwrap()]);
+        let err = run(args)?;
+
+        assert_eq!(err, ExitStatus::Success);
+        Ok(())
+    }
+
+    #[test]
+    fn test_check_returns_cleanly_for_multiline_strings_with_crlf_line_endings(
+    ) -> anyhow::Result<()> {
+        let fixtures = path_fixtures();
+        let path = fixtures.join("crlf").join("multiline_string_value.R");
+        let path = path.to_str().unwrap();
+
+        let args = Args::parse_from(["", "format", path, "--check"]);
+        let err = run(args)?;
+
+        assert_eq!(err, ExitStatus::Success);
+        Ok(())
+    }
+}
